move the cuindicator
