+++ modules/rapids_modules/analysis/outCsvNode.py
@@ -1,7 +1,7 @@
 from gquant.dataframe_flow import Node
 import dask_cudf
 from gquant.dataframe_flow.portsSpecSchema import ConfSchema
-from gquant.dataframe_flow._port_type_node import _PortTypesMixin
+from .._port_type_node import _PortTypesMixin
 
 
 class OutCsvNode(Node, _PortTypesMixin):
