+++ modules/rapids_modules/analysis/scatterPlotNode.py
@@ -4,7 +4,7 @@ from bqplot import (Axis, LinearScale,  Figure,
 import dask_cudf
 import cudf
 from gquant.dataframe_flow.portsSpecSchema import ConfSchema
-from gquant.dataframe_flow._port_type_node import _PortTypesMixin
+from .._port_type_node import _PortTypesMixin
 
 __all__ = ["ScatterPlotNode"]
