+++ modules/rapids_modules/analysis/sharpeRatioNode.py
@@ -2,7 +2,7 @@ from gquant.dataframe_flow import Node
 import math
 import dask_cudf
 from gquant.dataframe_flow.portsSpecSchema import ConfSchema
-from gquant.dataframe_flow._port_type_node import _PortTypesMixin
+from .._port_type_node import _PortTypesMixin
 
 
 class SharpeRatioNode(Node, _PortTypesMixin):
