+++ modules/rapids_modules/ml/forestInference.py
@@ -1,7 +1,7 @@
 from gquant.dataframe_flow import Node
 import cudf
 import dask_cudf
-from gquant.dataframe_flow._port_type_node import _PortTypesMixin
+from .._port_type_node import _PortTypesMixin
 from gquant.dataframe_flow.portsSpecSchema import (ConfSchema,
                                                    PortsSpecSchema, NodePorts)
 import copy
