+++ modules/rapids_modules/ml/splitDataNode.py
@@ -1,4 +1,4 @@
-from gquant.dataframe_flow._port_type_node import _PortTypesMixin
+from .._port_type_node import _PortTypesMixin
 from gquant.dataframe_flow.portsSpecSchema import (ConfSchema,
                                                    NodePorts, PortsSpecSchema)
 from gquant.dataframe_flow import Node
