+++ modules/rapids_modules/strategy/xgboostStrategyNode.py
@@ -4,7 +4,7 @@ import cudf
 import dask_cudf
 import xgboost as xgb
 import dask
-from gquant.dataframe_flow._port_type_node import _PortTypesMixin
+from .._port_type_node import _PortTypesMixin
 from gquant.dataframe_flow.portsSpecSchema import ConfSchema
