+++ modules/rapids_modules/transform/daskComputeNode.py
@@ -5,7 +5,7 @@ import dask_cudf
 
 from gquant.dataframe_flow import Node
 from gquant.dataframe_flow.portsSpecSchema import ConfSchema
-from gquant.dataframe_flow._port_type_node import _PortTypesMixin
+from .._port_type_node import _PortTypesMixin
 
 from gquant.dataframe_flow.portsSpecSchema import (
     PortsSpecSchema, NodePorts)
