+++ modules/rapids_modules/transform/dropNode.py
@@ -1,5 +1,5 @@
 from gquant.dataframe_flow import Node
-from gquant.dataframe_flow._port_type_node import _PortTypesMixin
+from .._port_type_node import _PortTypesMixin
 from gquant.dataframe_flow.portsSpecSchema import ConfSchema
