+++ modules/rapids_modules/transform/indicatorNode.py
@@ -1,5 +1,5 @@
-import gquant.cuindicator as ci
-from gquant.dataframe_flow._port_type_node import _PortTypesMixin
+from .. import cuindicator as ci
+from .._port_type_node import _PortTypesMixin
 from gquant.dataframe_flow.portsSpecSchema import ConfSchema
 from gquant.dataframe_flow import Node
 import copy
