+++ modules/rapids_modules/transform/onehotEncoding.py
@@ -1,6 +1,6 @@
 from gquant.dataframe_flow import Node
 from gquant.dataframe_flow.portsSpecSchema import ConfSchema
-from gquant.dataframe_flow._port_type_node import _PortTypesMixin
+from .._port_type_node import _PortTypesMixin
 
 __all__ = ["OneHotEncodingNode"]
